+++ src/main/java/insane96mcp/progressivebosses/ProgressiveBosses.java
@@ -23,7 +23,6 @@ import net.minecraftforge.event.entity.EntityAttributeCreationEvent;
 import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.config.ModConfig;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
@@ -39,10 +38,10 @@ public class ProgressiveBosses {
 
 	public static final Logger LOGGER = LogManager.getLogger();
 
-	public ProgressiveBosses() {
-		ModLoadingContext.get().registerConfig(ModConfig.Type.COMMON, Config.COMMON_SPEC);
+	public ProgressiveBosses(FMLJavaModLoadingContext modContext) {
+		modContext.registerConfig(ModConfig.Type.COMMON, Config.COMMON_SPEC);
 		MinecraftForge.EVENT_BUS.register(this);
-		final IEventBus modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
+		final IEventBus modEventBus = modContext.getModEventBus();
 		modEventBus.addListener(ClientSetup::registerEntityRenderers);
 		modEventBus.addListener(ClientSetup::creativeTabsBuildContents);
 		modEventBus.addListener(this::commonSetup);