+++ src/main/java/insane96mcp/progressivebosses/module/wither/WitherFeature.java
@@ -3,6 +3,7 @@ package insane96mcp.progressivebosses.module.wither;
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.LoadFeature;
 import insane96mcp.insanelib.base.Module;
+import insane96mcp.insanelib.base.config.Config;
 import insane96mcp.progressivebosses.ProgressiveBosses;
 import insane96mcp.progressivebosses.module.wither.entity.PBWither;
 import insane96mcp.progressivebosses.setup.PBEntities;
@@ -27,6 +28,9 @@ import java.util.List;
 public class WitherFeature extends Feature {
 	public static final TagKey<Item> WITHER_INVULNERABLE = ItemTags.create(ResourceLocation.fromNamespaceAndPath(ProgressiveBosses.MOD_ID, "wither_invulnerable"));
 	public static final TagKey<Item> WORLD_INVULNERABLE = ItemTags.create(ResourceLocation.fromNamespaceAndPath(ProgressiveBosses.MOD_ID, "world_invulnerable"));
+
+	@Config(description = "If true, Wither can charge any entity and not just players.")
+	public static Boolean allowChargingNonPlayers = false;
 
 	public WitherFeature(Module module, boolean enabledByDefault, boolean canBeDisabled) {
 		super(module, enabledByDefault, canBeDisabled);