+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Upcoming
+## 4.3.5
 * Added tooltip to Dragon Egg to let you know that more drop from dragons
 * Wither no longer charges non-players
   * Configurable